chore: restructure infocards
