+++ packages/components/src/templates/next/components/complex/InfoCards/components/index.ts
@@ -1,0 +1,5 @@
+export * from "./InfoCardContainer"
+export * from "./InfoCardImage"
+export * from "./InfoCardNoImage"
+export * from "./InfoCardWithImage"
+export * from "./InfoCardText"
